A visual program done in JavaFX as a test facility.
Note you need to download the jfx sdk depending on which  ide program your using.

It uses JavaFX to create a program window, creat a table that auto-adjusts thanks to the API built in to JavaFX. I classified four columns and added text fields for input to those columns, aswell as a save and delete button. Rows SHOULD be visually implemented into the table whenever the list is called.
*A lot of work is done by the JavaFX API so I call a number of there methods.

Problems - I cant seem to get the 'List' to show up on the 'Table' I set. I was hoping you could take a look at it and tell me what Im doing wrong. Because as far as errors, I dont get any.